pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "seed-agent"
version = "0.1.0"
description = "Deploy autonomous AI agents that grow through experience"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    {name = "Sean Florez-Humphrey", email = "sean@groundstate.ai"},
]
keywords = ["ai", "autonomous", "agent", "seed", "connectors"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "pyyaml>=6.0",
]

[project.optional-dependencies]
slack = ["slack-bolt>=1.18.0", "slack-sdk>=3.21.0"]
discord = ["discord.py>=2.3.0"]
telegram = ["python-telegram-bot>=20.0"]
webhook = ["flask>=3.0.0"]
all = ["seed-agent[slack,discord,telegram,webhook]"]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]

[project.urls]
Homepage = "https://github.com/sfhsdev/seed-agent"
Repository = "https://github.com/sfhsdev/seed-agent"
Documentation = "https://github.com/sfhsdev/seed-agent/tree/main/docs"

[tool.setuptools.packages.find]
include = ["connectors*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
